bcm53xx: remove MR32 platform, pre-upgrade checks
This patch removes use of the nand_do_platform_check() function
and replaces it with REQUIRE_IMAGE_METADATA=1 for the MR32, with
REQUIRE_IMAGE_METADATA explicitly 0 elsewhere. This is based upon
bcm63xx's base-files/lib/upgrade/platform.sh.

platform_nand_pre_upgrade() was removed in 2019 [1] and thus is not
needed.

[1]http://lists.openwrt.org/pipermail/openwrt-devel/2019-July/023719.html

Signed-off-by: Matthew Hagan <mnhagan88@gmail.com>